Camera and Jump
Added camera smoothing as well as setting limits for how far left, right, up, and down it can go. Adjusted the jump to have increased gravity while falling to aid in the jump feeling less 'floaty'. Added collision checkers in player to check if the player is on the ground so that the play can only jump when they are 'grounded'